% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pfam.R
\name{load_pfam}
\alias{load_pfam}
\title{load pfam}
\usage{
load_pfam()
}
\description{
a convenience function to load the PFAM data files and combine them together
into a single table. The resulting data.table is ~360Mb.
}

