% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFAMProteinData.R
\docType{data}
\name{PFAM_data2}
\alias{PFAM_data2}
\title{PFAM boundaries for all UniProt proteins: Part2}
\format{A  \code{\link[data.table]{data.table}} with 71810937 rows and 4 variables, with the key set on the PFAM_ID column:
\describe{
  \item{PFAM_ID}{PFAM unique identifier}
  \item{UNIPROT_ACC}{the accession of the UniProtID}
  \item{START}{start of the PFAM hmm match of the PFAM_ID against teh UNIPROT_ACC}
  \item{END}{end of the PFAM hmm match of the PFAM_ID against teh UNIPROT_ACC}
}}
\source{
\url{ftp://ftp.ebi.ac.uk/pub/databases/Pfam/current_release}
}
\usage{
PFAM_data2
}
\description{
This table defines the boundaries of all PFAM domains against all UniProt proteins. It is a direct derivative of the PFAM-A.full.uniprot.gz
file available on the PFAM ftp website.
}
\keyword{datasets}

