% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFAMProteinData.R
\docType{package}
\name{PFAMProteinData}
\alias{PFAMProteinData}
\alias{PFAMProteinData-package}
\title{PFAMProteinData: A package for computating the notorious bar statistic.}
\description{
The PFAMProteinData package provides a single table that defines the boundaries of each
UniProt protein with each PFAM domain, and also provides the start and end of each of the domains.
The code to generate this table can be found in the data-raw directory and the ultimate source of the
data is the PFAM-A.full.uniprot.gz file available at the PFAM ftp site.
}
\details{
The current release is derived from PFAM release 30.
}

